feat: 🎸 Create an asset namespace
- sdk.currentIdentity.reserveTicker -> sdk.assets.reserveTicker
- sdk.currentIdentity.createSecurityToken -> sdk.assets.createSecurityToken
- sdk.claimClassicTicker -> sdk.assets.claimClassicTicker
- sdk.isTickerAvailable -> sdk.assets.isTickerAvailable
- sdk.getTickerReservation -> sdk.assets.getTickerReservation
- sdk.getTickerReservations -> sdk.assets.getTickerReservations
- sdk.getSecurityToken -> sdk.assets.getSecurityToken
- sdk.getSecurityTokens -> sdk.assets.getSecurityToken

BREAKING CHANGE: 🧨 Moved methods
