+++ src/Polymesh.ts
@@ -13,17 +13,14 @@ import { satisfies } from 'semver';
 import { w3cwebsocket as W3CWebSocket } from 'websocket';
 import WebSocketAsPromised from 'websocket-as-promised';
 
+import { Assets } from '~/Assets';
 import {
   Account,
-  claimClassicTicker,
-  ClaimClassicTickerParams,
   Context,
   Identity,
   PolymeshError,
   registerIdentity,
   RegisterIdentityParams,
-  SecurityToken,
-  TickerReservation,
   transferPolyx,
   TransferPolyxParams,
 } from '~/internal';
@@ -37,20 +34,16 @@ import {
   NetworkProperties,
   ProcedureMethod,
   SubCallback,
-  TickerReservationStatus,
   UiKeyring,
   UnsubCallback,
 } from '~/types';
 import {
   moduleAddressToString,
   signerToString,
-  stringToIdentityId,
-  stringToTicker,
   textToString,
-  tickerToString,
   u32ToBigNumber,
 } from '~/utils/conversion';
-import { createProcedureMethod, getDid, isPrintableAscii } from '~/utils/internal';
+import { createProcedureMethod } from '~/utils/internal';
 
 import { Claims } from './Claims';
 import { CurrentIdentity } from './CurrentIdentity';
@@ -82,6 +75,10 @@ export class Polymesh {
   public middleware: Middleware;
   public settlements: Settlements;
   public currentIdentity: CurrentIdentity;
+  /**
+   * A set of methods for interacting with Assets
+   */
+  public assets: Assets;
 
   /**
    * @hidden
@@ -95,6 +92,7 @@ export class Polymesh {
     this.middleware = new Middleware(context);
     this.settlements = new Settlements(context);
     this.currentIdentity = new CurrentIdentity(context);
+    this.assets = new Assets(context);
 
     this.transferPolyx = createProcedureMethod(
       { getProcedureAndArgs: args => [transferPolyx, args] },
@@ -103,13 +101,6 @@ export class Polymesh {
 
     this.registerIdentity = createProcedureMethod(
       { getProcedureAndArgs: args => [registerIdentity, args] },
-      context
-    );
-
-    this.claimClassicTicker = createProcedureMethod(
-      {
-        getProcedureAndArgs: args => [claimClassicTicker, args],
-      },
       context
     );
   }
@@ -350,114 +341,6 @@ export class Polymesh {
   }
 
   /**
-   * Claim a ticker symbol that was reserved in Polymath Classic (Ethereum). The Ethereum account
-   *   that owns the ticker must sign a special message that contains the DID of the Identity that will own the ticker
-   *   in Polymesh, and provide the signed data to this call
-   */
-  public claimClassicTicker: ProcedureMethod<ClaimClassicTickerParams, TickerReservation>;
-
-  /**
-   * Check if a ticker hasn't been reserved
-   *
-   * @note can be subscribed to
-   */
-  public isTickerAvailable(args: { ticker: string }): Promise<boolean>;
-  public isTickerAvailable(
-    args: { ticker: string },
-    callback: SubCallback<boolean>
-  ): Promise<UnsubCallback>;
-
-  // eslint-disable-next-line require-jsdoc
-  public async isTickerAvailable(
-    args: { ticker: string },
-    callback?: SubCallback<boolean>
-  ): Promise<boolean | UnsubCallback> {
-    const reservation = new TickerReservation(args, this.context);
-
-    if (callback) {
-      return reservation.details(({ status: reservationStatus }) => {
-        // eslint-disable-next-line standard/no-callback-literal
-        callback(reservationStatus === TickerReservationStatus.Free);
-      });
-    }
-    const { status } = await reservation.details();
-
-    return status === TickerReservationStatus.Free;
-  }
-
-  /**
-   * Retrieve all the ticker reservations currently owned by an Identity. This doesn't include tokens that
-   *   have already been launched
-   *
-   * @param args.owner - defaults to the current Identity
-   *
-   * @note reservations with unreadable characters in their tickers will be left out
-   */
-  public async getTickerReservations(args?: {
-    owner: string | Identity;
-  }): Promise<TickerReservation[]> {
-    const {
-      context: {
-        polymeshApi: { query },
-      },
-      context,
-    } = this;
-
-    const did = await getDid(args?.owner, context);
-
-    const entries = await query.asset.assetOwnershipRelations.entries(
-      stringToIdentityId(did, context)
-    );
-
-    return entries.reduce<TickerReservation[]>((result, [key, relation]) => {
-      if (relation.isTickerOwned) {
-        const ticker = tickerToString(key.args[1]);
-
-        if (isPrintableAscii(ticker)) {
-          return [...result, new TickerReservation({ ticker }, context)];
-        }
-      }
-
-      return result;
-    }, []);
-  }
-
-  /**
-   * Retrieve a Ticker Reservation
-   *
-   * @param args.ticker - Security Token ticker
-   */
-  public async getTickerReservation(args: { ticker: string }): Promise<TickerReservation> {
-    const { ticker } = args;
-    const {
-      context: {
-        polymeshApi: {
-          query: { asset },
-        },
-      },
-      context,
-    } = this;
-
-    const { owner, expiry } = await asset.tickers(stringToTicker(ticker, context));
-
-    if (!owner.isEmpty) {
-      if (!expiry.isNone) {
-        return new TickerReservation({ ticker }, context);
-      }
-
-      throw new PolymeshError({
-        code: ErrorCode.UnmetPrerequisite,
-        message: `${ticker} token has been created`,
-      });
-    }
-
-    throw new PolymeshError({
-      code: ErrorCode.UnmetPrerequisite,
-      message: `There is no reservation for ${ticker} ticker`,
-    });
-  }
-
-  /**
    * Create an Identity instance from a DID
    *
    * @throws if there is no Identity with the passed DID
@@ -568,61 +451,6 @@ export class Polymesh {
     return (): void => {
       polymeshApi.off('disconnected', callback);
     };
-  }
-
-  /**
-   * Retrieve all the Security Tokens owned by an Identity
-   *
-   * @param args.owner - identity representation or Identity ID as stored in the blockchain
-   *
-   * @note tokens with unreadable characters in their tickers will be left out
-   */
-  public async getSecurityTokens(args?: { owner: string | Identity }): Promise<SecurityToken[]> {
-    const {
-      context: {
-        polymeshApi: { query },
-      },
-      context,
-    } = this;
-
-    const did = await getDid(args?.owner, context);
-
-    const entries = await query.asset.assetOwnershipRelations.entries(
-      stringToIdentityId(did, context)
-    );
-
-    return entries.reduce<SecurityToken[]>((result, [key, relation]) => {
-      if (relation.isAssetOwned) {
-        const ticker = tickerToString(key.args[1]);
-
-        if (isPrintableAscii(ticker)) {
-          return [...result, new SecurityToken({ ticker }, context)];
-        }
-      }
-
-      return result;
-    }, []);
-  }
-
-  /**
-   * Retrieve a Security Token
-   *
-   * @param args.ticker - Security Token ticker
-   */
-  public async getSecurityToken(args: { ticker: string }): Promise<SecurityToken> {
-    const { ticker } = args;
-
-    const token = new SecurityToken({ ticker }, this.context);
-    const exists = await token.exists();
-
-    if (!exists) {
-      throw new PolymeshError({
-        code: ErrorCode.DataUnavailable,
-        message: `There is no Security Token with ticker "${ticker}"`,
-      });
-    }
-
-    return token;
   }
 
   /**
