+++ src/__tests__/Assets.ts
@@ -1,0 +1,436 @@
+import BigNumber from 'bignumber.js';
+import sinon from 'sinon';
+
+import { Assets } from '~/Assets';
+import { Context, SecurityToken, TickerReservation, TransactionQueue } from '~/internal';
+import { dsMockUtils, entityMockUtils, procedureMockUtils } from '~/testUtils/mocks';
+import { Mocked } from '~/testUtils/types';
+import { KnownTokenType, TickerReservationStatus, TokenIdentifierType } from '~/types';
+import { tuple } from '~/types/utils';
+import * as utilsConversionModule from '~/utils/conversion';
+
+jest.mock(
+  '@polkadot/api',
+  require('~/testUtils/mocks/dataSources').mockPolkadotModule('@polkadot/api')
+);
+jest.mock(
+  '~/base/Context',
+  require('~/testUtils/mocks/dataSources').mockContextModule('~/base/Context')
+);
+jest.mock(
+  'apollo-client',
+  require('~/testUtils/mocks/dataSources').mockApolloModule('apollo-client')
+);
+jest.mock(
+  '~/api/entities/TickerReservation',
+  require('~/testUtils/mocks/entities').mockTickerReservationModule(
+    '~/api/entities/TickerReservation'
+  )
+);
+jest.mock(
+  '~/api/entities/Identity',
+  require('~/testUtils/mocks/entities').mockIdentityModule('~/api/entities/Identity')
+);
+jest.mock(
+  '~/api/entities/Account',
+  require('~/testUtils/mocks/entities').mockAccountModule('~/api/entities/Account')
+);
+jest.mock(
+  '~/base/Procedure',
+  require('~/testUtils/mocks/procedure').mockProcedureModule('~/base/Procedure')
+);
+jest.mock(
+  '~/api/entities/SecurityToken',
+  require('~/testUtils/mocks/entities').mockSecurityTokenModule('~/api/entities/SecurityToken')
+);
+jest.mock(
+  'websocket-as-promised',
+  require('~/testUtils/mocks/dataSources').mockWebSocketAsPromisedModule()
+);
+jest.mock(
+  '~/base/Procedure',
+  require('~/testUtils/mocks/procedure').mockProcedureModule('~/base/Procedure')
+);
+
+describe('Assets Class', () => {
+  let context: Mocked<Context>;
+  let assets: Assets;
+
+  beforeAll(() => {
+    dsMockUtils.initMocks();
+    entityMockUtils.initMocks();
+    procedureMockUtils.initMocks();
+  });
+
+  beforeEach(() => {
+    context = dsMockUtils.getContextInstance();
+    assets = new Assets(context);
+  });
+
+  afterEach(() => {
+    dsMockUtils.reset();
+    entityMockUtils.reset();
+    procedureMockUtils.reset();
+  });
+
+  afterAll(() => {
+    dsMockUtils.cleanup();
+    entityMockUtils.cleanup();
+    procedureMockUtils.cleanup();
+  });
+
+  describe('method: reserveTicker', () => {
+    test('should prepare the procedure with the correct arguments and context, and return the resulting transaction queue', async () => {
+      const args = {
+        ticker: 'SOMETICKER',
+      };
+
+      const expectedQueue = ('someQueue' as unknown) as TransactionQueue<TickerReservation>;
+
+      procedureMockUtils
+        .getPrepareStub()
+        .withArgs({ args, transformer: undefined }, context)
+        .resolves(expectedQueue);
+
+      const queue = await assets.reserveTicker(args);
+
+      expect(queue).toBe(expectedQueue);
+    });
+  });
+
+  describe('method: createToken', () => {
+    test('should prepare the procedure with the correct arguments and context, and return the resulting transaction queue', async () => {
+      const ticker = 'FAKETICKER';
+
+      const args = {
+        ticker,
+        name: 'TEST',
+        totalSupply: new BigNumber(100),
+        isDivisible: true,
+        tokenType: KnownTokenType.EquityCommon,
+        tokenIdentifiers: [{ type: TokenIdentifierType.Isin, value: '12345' }],
+        fundingRound: 'Series A',
+        requireInvestorUniqueness: false,
+        reservationRequired: false,
+      };
+
+      const expectedQueue = ('someQueue' as unknown) as TransactionQueue<SecurityToken>;
+
+      procedureMockUtils
+        .getPrepareStub()
+        .withArgs({ args, transformer: undefined }, context)
+        .resolves(expectedQueue);
+
+      const queue = await assets.createToken(args);
+
+      expect(queue).toBe(expectedQueue);
+    });
+  });
+
+  describe('method: claimClassicTicker', () => {
+    test('should prepare the procedure with the correct arguments and context, and return the resulting transaction queue', async () => {
+      const args = {
+        ticker: 'SOMETICKER',
+        ethereumSignature: 'someSig',
+      };
+
+      const expectedQueue = ('someQueue' as unknown) as TransactionQueue<TickerReservation>;
+
+      procedureMockUtils
+        .getPrepareStub()
+        .withArgs({ args, transformer: undefined }, context)
+        .resolves(expectedQueue);
+
+      const queue = await assets.claimClassicTicker(args);
+
+      expect(queue).toBe(expectedQueue);
+    });
+  });
+
+  describe('method: isTickerAvailable', () => {
+    beforeAll(() => {
+      entityMockUtils.initMocks();
+    });
+
+    afterEach(() => {
+      entityMockUtils.reset();
+    });
+
+    afterAll(() => {
+      entityMockUtils.cleanup();
+    });
+
+    test('should return true if ticker is available to reserve it', async () => {
+      entityMockUtils.getTickerReservationDetailsStub().resolves({
+        owner: entityMockUtils.getIdentityInstance(),
+        expiryDate: new Date(),
+        status: TickerReservationStatus.Free,
+      });
+
+      const isTickerAvailable = await assets.isTickerAvailable({ ticker: 'someTicker' });
+
+      expect(isTickerAvailable).toBeTruthy();
+    });
+
+    test('should return false if ticker is not available to reserve it', async () => {
+      entityMockUtils.getTickerReservationDetailsStub().resolves({
+        owner: entityMockUtils.getIdentityInstance(),
+        expiryDate: new Date(),
+        status: TickerReservationStatus.Reserved,
+      });
+
+      const isTickerAvailable = await assets.isTickerAvailable({ ticker: 'someTicker' });
+
+      expect(isTickerAvailable).toBeFalsy();
+    });
+
+    test('should allow subscription', async () => {
+      const unsubCallback = 'unsubCallBack';
+
+      entityMockUtils.getTickerReservationDetailsStub().callsFake(async cbFunc => {
+        cbFunc({
+          owner: entityMockUtils.getIdentityInstance(),
+          expiryDate: new Date(),
+          status: TickerReservationStatus.Free,
+        });
+
+        return unsubCallback;
+      });
+
+      const callback = sinon.stub();
+      const result = await assets.isTickerAvailable({ ticker: 'someTicker' }, callback);
+
+      expect(result).toBe(unsubCallback);
+      sinon.assert.calledWithExactly(callback, true);
+    });
+  });
+
+  describe('method: getTickerReservations', () => {
+    beforeAll(() => {
+      sinon.stub(utilsConversionModule, 'signerValueToSignatory');
+    });
+
+    afterAll(() => {
+      sinon.restore();
+    });
+
+    test('should return a list of ticker reservations if did parameter is set', async () => {
+      const fakeTicker = 'TEST';
+      const did = 'someDid';
+
+      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
+
+      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
+        entries: [
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
+            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
+          ),
+        ],
+      });
+
+      const tickerReservations = await assets.getTickerReservations({ owner: did });
+
+      expect(tickerReservations).toHaveLength(1);
+      expect(tickerReservations[0].ticker).toBe(fakeTicker);
+    });
+
+    test('should return a list of ticker reservations owned by the Identity', async () => {
+      const fakeTicker = 'TEST';
+      const did = 'someDid';
+
+      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
+
+      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
+        entries: [
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
+            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
+          ),
+        ],
+      });
+
+      const tickerReservations = await assets.getTickerReservations();
+
+      expect(tickerReservations).toHaveLength(1);
+      expect(tickerReservations[0].ticker).toBe(fakeTicker);
+    });
+
+    test('should filter out tickers with unreadable characters', async () => {
+      const fakeTicker = 'TEST';
+      const unreadableTicker = String.fromCharCode(65533);
+      const did = 'someDid';
+
+      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
+
+      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
+        entries: [
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
+            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
+          ),
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker('someTicker')],
+            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
+          ),
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(unreadableTicker)],
+            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
+          ),
+        ],
+      });
+
+      const tickerReservations = await assets.getTickerReservations();
+
+      expect(tickerReservations).toHaveLength(1);
+      expect(tickerReservations[0].ticker).toBe(fakeTicker);
+    });
+  });
+
+  describe('method: getTickerReservation', () => {
+    test('should return a specific ticker reservation owned by the Identity', async () => {
+      const ticker = 'TEST';
+      const expiry = new Date();
+
+      dsMockUtils.createQueryStub('asset', 'tickers', {
+        returnValue: dsMockUtils.createMockTickerRegistration({
+          owner: dsMockUtils.createMockIdentityId('someDid'),
+          expiry: dsMockUtils.createMockOption(dsMockUtils.createMockMoment(expiry.getTime())),
+        }),
+      });
+
+      const tickerReservation = await assets.getTickerReservation({ ticker });
+      expect(tickerReservation.ticker).toBe(ticker);
+    });
+
+    test('should throw if ticker reservation does not exist', async () => {
+      const ticker = 'TEST';
+
+      dsMockUtils.createQueryStub('asset', 'tickers', {
+        returnValue: dsMockUtils.createMockTickerRegistration({
+          owner: dsMockUtils.createMockIdentityId(),
+          expiry: dsMockUtils.createMockOption(),
+        }),
+      });
+
+      return expect(assets.getTickerReservation({ ticker })).rejects.toThrow(
+        `There is no reservation for ${ticker} ticker`
+      );
+    });
+
+    test('should throw if ticker is already a token', async () => {
+      const ticker = 'TEST';
+
+      dsMockUtils.createQueryStub('asset', 'tickers', {
+        returnValue: dsMockUtils.createMockTickerRegistration({
+          owner: dsMockUtils.createMockIdentityId('someDid'),
+          expiry: dsMockUtils.createMockOption(),
+        }),
+      });
+
+      return expect(assets.getTickerReservation({ ticker })).rejects.toThrow(
+        `${ticker} token has been created`
+      );
+    });
+  });
+
+  describe('method: getSecurityToken', () => {
+    test('should return a specific Security Token', async () => {
+      const ticker = 'TEST';
+
+      const securityToken = await assets.getSecurityToken({ ticker });
+      expect(securityToken.ticker).toBe(ticker);
+    });
+
+    test('should throw if the Security Token does not exist', async () => {
+      const ticker = 'TEST';
+      entityMockUtils.configureMocks({ securityTokenOptions: { exists: false } });
+
+      return expect(assets.getSecurityToken({ ticker })).rejects.toThrow(
+        `There is no Security Token with ticker "${ticker}"`
+      );
+    });
+  });
+
+  describe('method: getSecurityTokens', () => {
+    beforeAll(() => {
+      sinon.stub(utilsConversionModule, 'signerValueToSignatory');
+    });
+
+    afterAll(() => {
+      sinon.restore();
+    });
+
+    test('should return a list of security tokens owned by the supplied did', async () => {
+      const fakeTicker = 'TEST';
+      const did = 'someDid';
+
+      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
+
+      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
+        entries: [
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
+            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
+          ),
+        ],
+      });
+
+      const securityTokens = await assets.getSecurityTokens({ owner: 'someDid' });
+
+      expect(securityTokens).toHaveLength(1);
+      expect(securityTokens[0].ticker).toBe(fakeTicker);
+    });
+
+    test('should return a list of security tokens owned by the current Identity if no did is supplied', async () => {
+      const fakeTicker = 'TEST';
+      const did = 'someDid';
+
+      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
+
+      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
+        entries: [
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
+            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
+          ),
+        ],
+      });
+
+      const securityTokens = await assets.getSecurityTokens();
+
+      expect(securityTokens).toHaveLength(1);
+      expect(securityTokens[0].ticker).toBe(fakeTicker);
+    });
+
+    test('should filter out tokens whose tickers have unreadable characters', async () => {
+      const fakeTicker = 'TEST';
+      const unreadableTicker = String.fromCharCode(65533);
+      const did = 'someDid';
+
+      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
+
+      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
+        entries: [
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
+            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
+          ),
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker('someTicker')],
+            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
+          ),
+          tuple(
+            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(unreadableTicker)],
+            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
+          ),
+        ],
+      });
+
+      const securityTokens = await assets.getSecurityTokens();
+
+      expect(securityTokens).toHaveLength(1);
+      expect(securityTokens[0].ticker).toBe(fakeTicker);
+    });
+  });
+});
