+++ src/__tests__/Polymesh.ts
@@ -7,14 +7,12 @@ import { TxTags } from 'polymesh-types/types';
 import semver from 'semver';
 import sinon from 'sinon';
 
-import { Account, Identity, TickerReservation, TransactionQueue } from '~/internal';
+import { Account, Identity, TransactionQueue } from '~/internal';
 import { heartbeat } from '~/middleware/queries';
 import { Polymesh } from '~/Polymesh';
 import { dsMockUtils, entityMockUtils, procedureMockUtils } from '~/testUtils/mocks';
-import { AccountBalance, SubCallback, TickerReservationStatus } from '~/types';
-import { tuple } from '~/types/utils';
+import { AccountBalance, SubCallback } from '~/types';
 import { SUPPORTED_VERSION_RANGE } from '~/utils/constants';
-import * as utilsConversionModule from '~/utils/conversion';
 
 jest.mock(
   '@polkadot/api',
@@ -406,264 +404,6 @@ describe('Polymesh Class', () => {
     });
   });
 
-  describe('method: claimClassicTicker', () => {
-    test('should prepare the procedure with the correct arguments and context, and return the resulting transaction queue', async () => {
-      const context = dsMockUtils.getContextInstance();
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-      });
-
-      const args = {
-        ticker: 'SOMETICKER',
-        ethereumSignature: 'someSig',
-      };
-
-      const expectedQueue = ('someQueue' as unknown) as TransactionQueue<TickerReservation>;
-
-      procedureMockUtils
-        .getPrepareStub()
-        .withArgs({ args, transformer: undefined }, context)
-        .resolves(expectedQueue);
-
-      const queue = await polymesh.claimClassicTicker(args);
-
-      expect(queue).toBe(expectedQueue);
-    });
-  });
-
-  describe('method: isTickerAvailable', () => {
-    beforeAll(() => {
-      entityMockUtils.initMocks();
-    });
-
-    afterEach(() => {
-      entityMockUtils.reset();
-    });
-
-    afterAll(() => {
-      entityMockUtils.cleanup();
-    });
-
-    test('should return true if ticker is available to reserve it', async () => {
-      entityMockUtils.getTickerReservationDetailsStub().resolves({
-        owner: entityMockUtils.getIdentityInstance(),
-        expiryDate: new Date(),
-        status: TickerReservationStatus.Free,
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const isTickerAvailable = await polymesh.isTickerAvailable({ ticker: 'someTicker' });
-
-      expect(isTickerAvailable).toBeTruthy();
-    });
-
-    test('should return false if ticker is not available to reserve it', async () => {
-      entityMockUtils.getTickerReservationDetailsStub().resolves({
-        owner: entityMockUtils.getIdentityInstance(),
-        expiryDate: new Date(),
-        status: TickerReservationStatus.Reserved,
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const isTickerAvailable = await polymesh.isTickerAvailable({ ticker: 'someTicker' });
-
-      expect(isTickerAvailable).toBeFalsy();
-    });
-
-    test('should allow subscription', async () => {
-      const unsubCallback = 'unsubCallBack';
-
-      entityMockUtils.getTickerReservationDetailsStub().callsFake(async cbFunc => {
-        cbFunc({
-          owner: entityMockUtils.getIdentityInstance(),
-          expiryDate: new Date(),
-          status: TickerReservationStatus.Free,
-        });
-
-        return unsubCallback;
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const callback = sinon.stub();
-      const result = await polymesh.isTickerAvailable({ ticker: 'someTicker' }, callback);
-
-      expect(result).toBe(unsubCallback);
-      sinon.assert.calledWithExactly(callback, true);
-    });
-  });
-
-  describe('method: getTickerReservations', () => {
-    beforeAll(() => {
-      sinon.stub(utilsConversionModule, 'signerValueToSignatory');
-    });
-
-    afterAll(() => {
-      sinon.restore();
-    });
-
-    test('should return a list of ticker reservations if did parameter is set', async () => {
-      const fakeTicker = 'TEST';
-      const did = 'someDid';
-
-      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
-
-      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
-        entries: [
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
-            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
-          ),
-        ],
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const tickerReservations = await polymesh.getTickerReservations({ owner: did });
-
-      expect(tickerReservations).toHaveLength(1);
-      expect(tickerReservations[0].ticker).toBe(fakeTicker);
-    });
-
-    test('should return a list of ticker reservations owned by the Identity', async () => {
-      const fakeTicker = 'TEST';
-      const did = 'someDid';
-
-      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
-
-      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
-        entries: [
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
-            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
-          ),
-        ],
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const tickerReservations = await polymesh.getTickerReservations();
-
-      expect(tickerReservations).toHaveLength(1);
-      expect(tickerReservations[0].ticker).toBe(fakeTicker);
-    });
-
-    test('should filter out tickers with unreadable characters', async () => {
-      const fakeTicker = 'TEST';
-      const unreadableTicker = String.fromCharCode(65533);
-      const did = 'someDid';
-
-      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
-
-      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
-        entries: [
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
-            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
-          ),
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker('someTicker')],
-            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
-          ),
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(unreadableTicker)],
-            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
-          ),
-        ],
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const tickerReservations = await polymesh.getTickerReservations();
-
-      expect(tickerReservations).toHaveLength(1);
-      expect(tickerReservations[0].ticker).toBe(fakeTicker);
-    });
-  });
-
-  describe('method: getTickerReservation', () => {
-    test('should return a specific ticker reservation owned by the Identity', async () => {
-      const ticker = 'TEST';
-      const expiry = new Date();
-
-      dsMockUtils.createQueryStub('asset', 'tickers', {
-        returnValue: dsMockUtils.createMockTickerRegistration({
-          owner: dsMockUtils.createMockIdentityId('someDid'),
-          expiry: dsMockUtils.createMockOption(dsMockUtils.createMockMoment(expiry.getTime())),
-        }),
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const tickerReservation = await polymesh.getTickerReservation({ ticker });
-      expect(tickerReservation.ticker).toBe(ticker);
-    });
-
-    test('should throw if ticker reservation does not exist', async () => {
-      const ticker = 'TEST';
-
-      dsMockUtils.createQueryStub('asset', 'tickers', {
-        returnValue: dsMockUtils.createMockTickerRegistration({
-          owner: dsMockUtils.createMockIdentityId(),
-          expiry: dsMockUtils.createMockOption(),
-        }),
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      return expect(polymesh.getTickerReservation({ ticker })).rejects.toThrow(
-        `There is no reservation for ${ticker} ticker`
-      );
-    });
-
-    test('should throw if ticker is already a token', async () => {
-      const ticker = 'TEST';
-
-      dsMockUtils.createQueryStub('asset', 'tickers', {
-        returnValue: dsMockUtils.createMockTickerRegistration({
-          owner: dsMockUtils.createMockIdentityId('someDid'),
-          expiry: dsMockUtils.createMockOption(),
-        }),
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      return expect(polymesh.getTickerReservation({ ticker })).rejects.toThrow(
-        `${ticker} token has been created`
-      );
-    });
-  });
-
   describe('method: getIdentity', () => {
     test('should return an Identity object with the passed did', async () => {
       const polymesh = await Polymesh.connect({
@@ -814,103 +554,6 @@ describe('Polymesh Class', () => {
     });
   });
 
-  describe('method: getSecurityTokens', () => {
-    beforeAll(() => {
-      sinon.stub(utilsConversionModule, 'signerValueToSignatory');
-    });
-
-    afterAll(() => {
-      sinon.restore();
-    });
-
-    test('should return a list of security tokens owned by the supplied did', async () => {
-      const fakeTicker = 'TEST';
-      const did = 'someDid';
-
-      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
-
-      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
-        entries: [
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
-            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
-          ),
-        ],
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const securityTokens = await polymesh.getSecurityTokens({ owner: 'someDid' });
-
-      expect(securityTokens).toHaveLength(1);
-      expect(securityTokens[0].ticker).toBe(fakeTicker);
-    });
-
-    test('should return a list of security tokens owned by the current Identity if no did is supplied', async () => {
-      const fakeTicker = 'TEST';
-      const did = 'someDid';
-
-      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
-
-      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
-        entries: [
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
-            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
-          ),
-        ],
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const securityTokens = await polymesh.getSecurityTokens();
-
-      expect(securityTokens).toHaveLength(1);
-      expect(securityTokens[0].ticker).toBe(fakeTicker);
-    });
-
-    test('should filter out tokens whose tickers have unreadable characters', async () => {
-      const fakeTicker = 'TEST';
-      const unreadableTicker = String.fromCharCode(65533);
-      const did = 'someDid';
-
-      dsMockUtils.configureMocks({ contextOptions: { withSeed: true } });
-
-      dsMockUtils.createQueryStub('asset', 'assetOwnershipRelations', {
-        entries: [
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(fakeTicker)],
-            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
-          ),
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker('someTicker')],
-            dsMockUtils.createMockAssetOwnershipRelation('TickerOwned')
-          ),
-          tuple(
-            [dsMockUtils.createMockIdentityId(did), dsMockUtils.createMockTicker(unreadableTicker)],
-            dsMockUtils.createMockAssetOwnershipRelation('AssetOwned')
-          ),
-        ],
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const securityTokens = await polymesh.getSecurityTokens();
-
-      expect(securityTokens).toHaveLength(1);
-      expect(securityTokens[0].ticker).toBe(fakeTicker);
-    });
-  });
-
   describe('method: transferPolyx', () => {
     test('should prepare the procedure with the correct arguments and context, and return the resulting transaction queue', async () => {
       const context = dsMockUtils.getContextInstance();
@@ -935,34 +578,6 @@ describe('Polymesh Class', () => {
       const queue = await polymesh.transferPolyx(args);
 
       expect(queue).toBe(expectedQueue);
-    });
-  });
-
-  describe('method: getSecurityToken', () => {
-    test('should return a specific Security Token', async () => {
-      const ticker = 'TEST';
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const securityToken = await polymesh.getSecurityToken({ ticker });
-      expect(securityToken.ticker).toBe(ticker);
-    });
-
-    test('should throw if the Security Token does not exist', async () => {
-      const ticker = 'TEST';
-      entityMockUtils.configureMocks({ securityTokenOptions: { exists: false } });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      return expect(polymesh.getSecurityToken({ ticker })).rejects.toThrow(
-        `There is no Security Token with ticker "${ticker}"`
-      );
     });
   });
